Slight Refactor
Since my plan is to forward packets across to other members, I didn't
want to decode (i.e convert from AppleLossless) only decrypt it.  So
that we still have an efficiently packed data packet, but the other
downstream members can save a bit of time and not have to decrypt the
packet.  Long story short, I want to forward decrypted but not decoded
packets.