Move to Go Modules (#7)
Moves to go modules and go 1.12. Ditching vendoring for now. Things seem to be all working. Needed a workaround for the arm build, I pull the dependencies into vendor and then the arm build will use the vendor to build, when building on arm I kept getting panics when it was resolving over the network so this my workaround for now.