Major Update (#9)
* start of adding in more fine control; intro of grpc for API

* Volume Control

First pass at volume control.  For forwarding player, the volume
received by the forwarding playing node will pass it on down stream,
so all nodes will currently have the same volume settings.

* Architecture Experimentation

Putting the pieces in place for one approach on how I am imaging the
architecture of a full deployment working.

* build fix

* adding first pass at mgmt server, as well as move to config file based config

* refactor to share cluster searching

* Vertical Slice Web Through Mgmt API

First big vertical slice of functionality.  Basic sanity of a web grpc
call flowing through hardcoded envoy, into a mgmt node that dynamically
joined the cluster.

Also got the stub in place for the frontend go portion to setup
for dynamically updating envoy

* Envoy Control Plane Integration

bcg-frontend will now be able to dynamically update its 'sidecar' proxy
with the addresses of bcg-mgmt API servers.  Right now, it will only
update to ones that are know on start up.  Dynamic addition and removal
coming next.

* adding dynamic adding and removing of nodes to Envoy cluster

* test of removing proto gen from before_install

* adding back gen to before install, remove from build

* renaming mgmt API object

* refactor to use a service layer

* First Pass of Backing Store

Adding a raft based store will hold higher level configuration
across our cluster of players.  First pass is API to hold
the display name

* Non Leader Forwarding

When a request comes in for a write to a non-leader node, we
will now look up and forward it to the leader

Not quite happy with the structure, need to find a way to
structure things to test /raft package

* only bootstrap when there only one node

* setting up frontend web code, boilerplate from: https://github.com/xxczaki/styled-react-boilerplate

* Web Client Integration

Getting a basic integration, with a bit of styling.  Integrates
with the get speakers grpc API and will render the active speakers
into a list

* temp commit proto for mgmt

* documentation

* solution for building everything for ARM

* output a little message for the build

* Move to Go Modules (#7)

Moves to go modules and go 1.12. Ditching vendoring for now. Things seem to be all working. Needed a workaround for the arm build, I pull the dependencies into vendor and then the arm build will use the vendor to build, when building on arm I kept getting panics when it was resolving over the network so this my workaround for now.

* updating gitignre; making proto script more useable

* adding persistence layer and grpc API for zones

* trying to add some stability. Added handling of the flush command

* Attempting to handle switching sources more gracefully

- Switched to grab a random data port for receiving
- Added dacp functionality to stop the stream from server side

* adding some unit tests for dacp

* using forwarding player always

* adding API to bcg node for adding/removing forwarding sessions

* slight refactor, using one constant for serviceType

* pushing name change down to speaker.  Starting zone adding

* zone creation end to end complete working

* changing service name back

* adding full crud APIs for zone mgmt

* fix error checking

* Adding Support For Packr

Can now package the web ui files into the go frontend binary,
using packr.  Did not update the build scripts yet, so the steps are
manual:

1. cd cmd/frontend/webui
2. npm install
3. npm run build
4. cd cmd/frontend
5  packr2
6  go build

* adding a simple component for zones, fixing zone delete API

* UI tweaks

* updating travis build to build full frontend binary using packr2

* updating generation path

* adding packr install to build script

* capturing the current track and passing it all the way up to the UI

* capturing the current track and passing it all the way up to the UI

* will now forward all track info down stream

* updated web ui

* make changing broadcast name explict

* missing dep

* UI tweaks

* attempting dev publishing to google storage

* favicon generation

* Update build.sh

* debug logs

* more build debugging

* more tests

* more testing

* removing some failures

* Docker front test (#8)

* test

* binary for frontend now works properly

* ci/cd build for frontend will be linux only for now

* reverting a change

* tester sudo

* more tests

* updating manifest for web ui

* starting to add documentation

* Closeable Client

Create a quick client that can be closed.  Because of the way
I built the service where it creates a new client to the speaker
for every incoming mgmt request, a new connection is created.
I thought it would be good for the connection to be closed everytime.

* make sure we remove speakers from other others when creating a new zone

* update docs